pyproject.toml
[project]
name = "toss-next-challenge"
version = "0.1.0"
description = ""
authors = [
    {name = "ds-wook",email = "leewook94@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "category-encoders (>=2.8.1,<3.0.0)",
    "numpy (>=1.18.5,<2.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "isort (>=6.0.1,<7.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "catboost (>=1.2.8,<2.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "polars (>=1.33.0,<2.0.0)",
    "prettytable (>=3.16.0,<4.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "torch (==2.3.1)",
    "torchvision (==0.18.1)",
    "torchaudio (==2.3.1)",
    "wandb (==0.17.2)",
    "xgboost (==2.0.1)",
    "tqdm (>=4.67.1,<5.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "deepctr-torch (>=0.2.9,<0.3.0)",
    "easydict (>=1.13,<2.0)",
    "gensim (>=4.3.3,<5.0.0)",
    "line-profiler (>=5.0.0,<6.0.0)",
]

[tool.poetry.group.dev.dependencies]
pre-commit = "4.0.*"
black = ">=24.10.0"
flake8 = ">=7.1.1"
isort = ">=5.13.2"
flake8-pyproject = ">=1.2.3"
pytest = "8.3.4"
pytest-dotenv = "^0.5.2"
ruff = "0.9.9"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Refer https://github.com/astral-sh/ruff?tab=readme-ov-file#configuration
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "notebook"
]

# Same as Black.
line-length = 88
indent-width = 4
